fix: seggregate logic for padre and trading-terminal
