+++ fees/padre/index.ts
@@ -7,12 +7,6 @@ import { CHAIN } from "../../helpers/chains";
 import { queryDuneSql } from "../../helpers/dune";
 const dataAvaliableTill = (Date.now() / 1e3 - 10 * 3600) // 10 hours ago
 const TRADING_TERMINAL_FEES = 'Trading Terminal Fees'
-const CASHBACK_REFERRAL_PAYOUTS = 'Cashback/Referral Payouts'
-const solanaConfig = {
-  mainFeeWallet: 'J5XGHmzrRmnYWbmw45DbYkdZAU2bwERFZ11qCDXPvFB5',
-  feeCashbackWallet: 'DoAsxPQgiyAxyaJNvpAAUb2ups6rbJRdYrCPyWxwRxBb',
-  burnWallet: '9jHrTCwpDANHLNQz5cem6XLUBM8KiTWKe766Br6KVCXM',
-}
 
 const evmChainConfig: any = {
   [CHAIN.ETHEREUM]: {
@@ -27,23 +21,8 @@ const evmChainConfig: any = {
 }
 
 async function fetchSolana(options: FetchOptions) {
-  const { mainFeeWallet, feeCashbackWallet, burnWallet } = solanaConfig
   const query = `
     WITH
-    interfundTransfers AS (
-      SELECT
-        tx_id
-      FROM
-        solana.account_activity
-      WHERE
-        TIME_RANGE
-        AND tx_success
-        AND address IN ('${mainFeeWallet}', '${feeCashbackWallet}')
-      GROUP BY tx_id
-      HAVING
-        SUM(CASE WHEN address = '${mainFeeWallet}' AND balance_change < 0 THEN 1 ELSE 0 END) > 0
-        AND SUM(CASE WHEN address = '${feeCashbackWallet}' AND balance_change > 0 THEN 1 ELSE 0 END) > 0
-    ),
     allFeePayments AS (
       SELECT
         tx_id,
@@ -52,30 +31,7 @@ async function fetchSolana(options: FetchOptions) {
         solana.account_activity
       WHERE
         TIME_RANGE
-        AND address IN ('${mainFeeWallet}', '${feeCashbackWallet}')
-        AND tx_success
-        AND balance_change > 0
-        AND NOT (address = '${feeCashbackWallet}' AND tx_id IN (SELECT tx_id FROM interfundTransfers))
-    ),
-    cashbackPayouts AS (
-      SELECT
-        COALESCE(SUM(-balance_change), 0) AS payout_amount
-      FROM
-        solana.account_activity
-      WHERE
-        TIME_RANGE
-        AND address = '${feeCashbackWallet}'
-        AND tx_success
-        AND balance_change < 0
-    ),
-    burnWalletInflows AS (
-      SELECT
-        COALESCE(SUM(balance_change), 0) AS buyback_burn_amount
-      FROM
-        solana.account_activity
-      WHERE
-        TIME_RANGE
-        AND address = '${burnWallet}'
+        AND address IN ('J5XGHmzrRmnYWbmw45DbYkdZAU2bwERFZ11qCDXPvFB5', 'DoAsxPQgiyAxyaJNvpAAUb2ups6rbJRdYrCPyWxwRxBb')
         AND tx_success
         AND balance_change > 0
     ),
@@ -88,33 +44,20 @@ async function fetchSolana(options: FetchOptions) {
         JOIN allFeePayments AS feePayments ON trades.tx_id = feePayments.tx_id
       WHERE
         TIME_RANGE
-        AND trades.trader_id != '${mainFeeWallet}'
-        AND trades.trader_id != '${feeCashbackWallet}'
-        AND trades.trader_id != '${burnWallet}'
+        AND trades.trader_id != 'J5XGHmzrRmnYWbmw45DbYkdZAU2bwERFZ11qCDXPvFB5'
+        AND trades.trader_id != 'DoAsxPQgiyAxyaJNvpAAUb2ups6rbJRdYrCPyWxwRxBb'
       GROUP BY trades.tx_id
     )
     SELECT
-      SUM(fee) AS fee,
-      (SELECT payout_amount FROM cashbackPayouts) AS cashback_payout_amount,
-      (SELECT buyback_burn_amount FROM burnWalletInflows) AS buyback_burn_amount
+      SUM(fee) AS fee
     FROM
       botTrades
   `;
 
-  const [row] = await queryDuneSql(options, query);
-  const tradingFees = Number(row?.fee ?? 0);
-  const cashbackPayouts = Number(row?.cashback_payout_amount ?? 0);
-  const buybackBurnAmount = Number(row?.buyback_burn_amount ?? 0);
-
+  const fees = await queryDuneSql(options, query);
   const dailyFees = options.createBalances();
-  const dailySupplySideRevenue = options.createBalances();
-  const dailyRevenue = options.createBalances();
-
-  dailyFees.add(ADDRESSES.solana.SOL, tradingFees, TRADING_TERMINAL_FEES);
-  dailySupplySideRevenue.add(ADDRESSES.solana.SOL, cashbackPayouts, CASHBACK_REFERRAL_PAYOUTS);
-  dailyRevenue.add(ADDRESSES.solana.SOL, tradingFees - cashbackPayouts - buybackBurnAmount, TRADING_TERMINAL_FEES);
-
-  return { dailyFees, dailyRevenue, dailyProtocolRevenue: dailyRevenue, dailySupplySideRevenue }
+  dailyFees.add(ADDRESSES.solana.SOL, fees[0].fee || 0);
+  return dailyFees
 }
 
 async function fetchEvm(options: FetchOptions) {
@@ -136,9 +79,8 @@ export const fetch = async (_: any, _1: any, options: FetchOptions) => {
   if (options.endTimestamp > dataAvaliableTill)
     throw new Error("Data not available till 10 hours ago. Please try a date before: " + new Date(dataAvaliableTill * 1e3).toISOString());
 
-  if (options.chain === CHAIN.SOLANA) return fetchSolana(options)
+  const fees = options.chain === CHAIN.SOLANA ? await fetchSolana(options) : await fetchEvm(options)
 
-  const fees = await fetchEvm(options)
   const dailyFees = fees.clone(1, TRADING_TERMINAL_FEES)
 
   return { dailyFees, dailyRevenue: dailyFees }
@@ -149,13 +91,7 @@ export const breakdownMethodology = {
     [TRADING_TERMINAL_FEES]: 'Fees charged on each trade executed through the trading terminal.',
   },
   Revenue: {
-    [TRADING_TERMINAL_FEES]: 'Trading terminal fees retained by the protocol after cashback/referral payouts and buyback/burn allocations.',
-  },
-  ProtocolRevenue: {
-    [TRADING_TERMINAL_FEES]: 'Trading terminal fees retained by the protocol after cashback/referral payouts and buyback/burn allocations.',
-  },
-  SupplySideRevenue: {
-    [CASHBACK_REFERRAL_PAYOUTS]: 'All outbound transfers from the cashback/referral wallet.',
+    [TRADING_TERMINAL_FEES]: 'Trading terminal fees retained by the protocol.',
   },
 }
 
@@ -170,11 +106,8 @@ const adapter: SimpleAdapter = {
   breakdownMethodology,
   methodology: {
     Fees: "Trading fees paid by users while using Padre bot.",
-    Revenue: "Trading terminal fees retained by the protocol after cashback/referral payouts and buyback/burn allocations.",
-    ProtocolRevenue: "Trading terminal fees retained by the protocol after cashback/referral payouts and buyback/burn allocations.",
-    SupplySideRevenue: "All outbound transfers from the cashback/referral wallet.",
+    Revenue: "All fees are collected by Padre protocol.",
   },
-  allowNegativeValue: true,
 };
 
 export default adapter;
